refactor(wallet): Move bump fee + foreign utxo tests
These files added to wallet/tests directory
- add_foreign_utxo.rs
- build_fee_bump.rs
